Creation of functional component and update of menu props
