+++ packages/web-components/src/components/cbp-menu/cbp-menu.stories.tsx
@@ -7,7 +7,11 @@ export default {
   argTypes: {
     position: {
       control: 'select',
-      options: [ 'bottom-start', "bottom-end", 'top-start', "top-end"]
+      options: [ 'bottom', "top"]
+    },
+    placement: {
+      control: 'select',
+      options: [ 'start', 'end']
     },
     context : {
       control: 'select',
@@ -36,11 +40,12 @@ function generateMenuItem(items){
   return html.join('');
 }
 
-const Template = ({ position, context, sx }) => {
+const Template = ({ position, placement, context, sx }) => {
   return ` 
     <cbp-menu
       uid="menuId"
       ${position ? `position="${position}"` : ''}
+      ${placement ? `placement="${placement}"` : ''}
       ${context ? `context="${context}"` : ''}
       ${sx ? `sx='${JSON.stringify(sx)}'` : ''}
     >
@@ -75,11 +80,12 @@ Menu.args = {};
 
 
 
-const ActionsMenuTemplate = ({ position, items, context, sx }) => {
+const ActionsMenuTemplate = ({ position, placement, items, context, sx }) => {
   return ` 
     <cbp-menu
       uid="menuId"
       ${position ? `position="${position}"` : ''}
+      ${placement ? `placement="${placement}"` : ''}
       ${context ? `context="${context}"` : ''}
       ${sx ? `sx='${JSON.stringify(sx)}'` : ''}
     >
